fix(sim): validate NPC role at save load + sign-compare in claim handler (#494)
Two production crashes from corrupt-state load paths.

1. server/sim_save.c::read_npc — an active NPC slot with a garbage
   role byte (saw 0x0B000007 in the wild) crashloop'd the server on
   first sim step: the despawn check fired with an out-of-range role,
   then character_free_for_npc dereferenced via that role and crashed.
   Validate after the full record is read so the file pointer is
   always past this NPC's bytes; on bad role just zero the slot and
   let the spawn loop refill it.

2. server/main.c claim_legacy_save handler — gcc on glibc rejected
   `hex_len + 1 > sizeof(basename)` as -Werror=sign-compare (uint8_t
   promotes to int, sizeof returns size_t). Cast to size_t. macOS
   apple-libc's looser warning hid this locally; only the docker
   server build on Linux flags it.

+++ server/main.c
@@ -666,7 +666,7 @@ static void handle_ws_message(struct mg_connection *c, struct mg_ws_message *wm)
          * "player_<slot>" form (anonymous slot fallback). Accept either:
          * try the literal name first, then with the historical prefix. */
         char basename[80];
-        if (hex_len + 1 > sizeof(basename)) goto claim_done;
+        if ((size_t)hex_len + 1 > sizeof(basename)) goto claim_done;
         memcpy(basename, hex, hex_len);
         basename[hex_len] = '\0';
